HABTM: User/Category
This relationship makes it easier to migrate the legacy data over. And it's
probably smart to keep the information around anyways.

For now, I've just added methods to the User model that satisfy the likely
calls that exist in the codebase elsewhere, so User#category and
User#category_id are still valid methods.